Fix Intel compiler warnings
Fallout from previous patch: Intel compiler
is very noisy.

No functional change.

+++ src/search.cpp
@@ -146,8 +146,8 @@ void Search::init() {
   // Init futility move count array
   for (d = 0; d < 32; ++d)
   {
-      FutilityMoveCounts[0][d] = 2.4 + 0.222 * pow(d + 0.00, 1.8);
-      FutilityMoveCounts[1][d] = 3.0 + 0.300 * pow(d + 0.98, 1.8);
+      FutilityMoveCounts[0][d] = int(2.4 + 0.222 * pow(d + 0.00, 1.8));
+      FutilityMoveCounts[1][d] = int(3.0 + 0.300 * pow(d + 0.98, 1.8));
   }
 }
